:recycle: Split util into conv, entropy, and log to make linter happy
