+++ internal/util/conv/conv.go
@@ -1,4 +1,4 @@
-package util
+package conv
 
 import "encoding/binary"
